Created module middleware, moved functions from container module
