+++ pilot/control/data.py
@@ -31,8 +31,9 @@ from pilot.util.common import should_abort
 from pilot.util.config import config
 from pilot.util.constants import PILOT_PRE_STAGEIN, PILOT_POST_STAGEIN, PILOT_PRE_STAGEOUT, PILOT_POST_STAGEOUT, LOG_TRANSFER_IN_PROGRESS,\
     LOG_TRANSFER_DONE, LOG_TRANSFER_NOT_DONE, LOG_TRANSFER_FAILED, SERVER_UPDATE_RUNNING, MAX_KILL_WAIT_TIME
-from pilot.util.container import execute, containerise_middleware
+from pilot.util.container import execute
 from pilot.util.filehandling import remove
+from pilot.util.middleware import containerise_middleware
 from pilot.util.processes import threads_aborted
 from pilot.util.queuehandling import declare_failed_by_kill, put_in_queue
 from pilot.util.timing import add_to_pilot_timing
@@ -185,10 +186,10 @@ def _stage_in(args, job):
         try:
             eventtype, localsite, remotesite = get_trace_report_variables(job)
             containerise_middleware(job, args.queue, script, eventtype, localsite, remotesite, stagein=True)
-        except Exception as e:
-            logger.warning('stage-in containerisation threw an exception: %s' % e)
         except PilotException as e:
             logger.warning('stage-in containerisation threw a pilot exception: %s' % e)
+        except Exception as e:
+            logger.warning('stage-in containerisation threw an exception: %s' % e)
     else:
         try:
             # create the trace report
